Flight status
status would be much more convenient and efficient way of showing that the plane is delayed/early/cancelled, etc.
I haven't included or working on the set function for status since we maybe could let the function detect the status on its own for example, if actualDeparture>scheduledDeparture the status can be delayed. Unsure about this hence waiting for your opinion.